requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
bleach==6.2.0
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2.23
django-allauth==0.57.2
django-summernote==0.8.20.0
django_celery_results==2.6.0
djangorestframework==3.16.0
gunicorn==20.1.0
idna==3.10
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==5.3.0
